# docs/api-spec.yml
openapi: 3.0.0
info:
  title: data-viewer API
  description: >
    Internal API served by the Next.js backend.
    Source API spec will be added once the external endpoint is defined.
  version: 0.5.0

paths:
  /api/auth/login:
    post:
      summary: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: Login successful — returns JWT token and user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Missing email or password, or invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/me:
    get:
      summary: Get current authenticated user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Missing, invalid, or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User no longer exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users:
    get:
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserProfile'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Non-admin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          description: Missing required fields, invalid role, or missing categoryId for player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Non-admin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/categories:
    get:
      summary: List all player categories
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesResponse'

  /api/fixture/matches:
    get:
      summary: List all tournament fixture matches
      description: >
        Proxies the Hockey Chubut external API, normalizes match data,
        and returns matches with status, score, venue, and round number.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixtureMatchesResponse'
        '500':
          description: External API failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/fixture/clubs:
    get:
      summary: List all clubs with logos
      description: >
        Proxies the Hockey Chubut external API and returns clubs
        with base64-encoded logo images.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixtureClubsResponse'
        '500':
          description: External API failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/players:
    get:
      summary: List players by category
      parameters:
        - name: categoryId
          in: query
          required: true
          schema:
            type: string
          description: ID of the category to filter players by
      responses:
        '200':
          description: Players for the given category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayersResponse'
        '400':
          description: Missing categoryId query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/fees:
    get:
      summary: Get current week's fees (role-filtered)
      description: >
        Admin sees all categories. Captain and player see only their assigned category.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fee configurations for the current week
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryFeeWithPlayers'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create or update fee configuration for a category (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryFeeRequest'
      responses:
        '201':
          description: Fee configuration created or updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  fee:
                    $ref: '#/components/schemas/CategoryFeeWithPlayers'
        '400':
          description: Missing required fields or non-positive values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Non-admin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/fees/mark-paid:
    post:
      summary: Mark a player's fee as paid (admin or captain)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                playerFeeId:
                  type: string
                  format: uuid
              required: [playerFeeId]
      responses:
        '200':
          description: Player fee marked as paid
          content:
            application/json:
              schema:
                type: object
                properties:
                  playerFee:
                    $ref: '#/components/schemas/PlayerFee'
        '400':
          description: Missing playerFeeId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Player role (only admin/captain allowed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Player fee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Player:
      type: object
      properties:
        id:
          type: string
        number:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
          enum: [active, inactive]
        categoryId:
          type: string
      required: [id, number, firstName, lastName, status, categoryId]

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]

    PlayersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Player'
        category:
          $ref: '#/components/schemas/Category'

    CategoriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'

    FixtureMatch:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [completed, pending]
        date:
          type: string
          format: date-time
        venue:
          type: string
        round:
          type: integer
        homeTeam:
          $ref: '#/components/schemas/TeamInfo'
        awayTeam:
          $ref: '#/components/schemas/TeamInfo'
        score:
          oneOf:
            - $ref: '#/components/schemas/MatchScore'
            - type: 'null'
      required: [id, status, date, venue, round, homeTeam, awayTeam, score]

    TeamInfo:
      type: object
      properties:
        clubId:
          type: integer
        clubName:
          type: string
      required: [clubId, clubName]

    MatchScore:
      type: object
      properties:
        home:
          type: integer
        away:
          type: integer
      required: [home, away]

    FixtureClub:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logo:
          type: string
          nullable: true
          description: Base64-encoded PNG image
      required: [id, name, logo]

    FixtureMatchesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FixtureMatch'

    FixtureClubsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FixtureClub'

    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, player, captain]
        firstName:
          type: string
        lastName:
          type: string
        categoryId:
          type: string
          nullable: true
      required: [id, email, role, firstName, lastName, categoryId]

    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, player, captain]
        firstName:
          type: string
        lastName:
          type: string
        categoryId:
          type: string
          nullable: true
          description: Required when role is 'player' or 'captain'
      required: [email, password, role, firstName, lastName]

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token (8h expiry)
        user:
          $ref: '#/components/schemas/UserProfile'
      required: [token, user]

    CategoryFeeWithPlayers:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
        categoryName:
          type: string
        totalAmount:
          type: number
        availablePlayers:
          type: integer
        perPlayerAmount:
          type: number
        weekStartDate:
          type: string
          format: date
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        playerFees:
          type: array
          items:
            $ref: '#/components/schemas/PlayerFee'
        paidCount:
          type: integer
        unpaidCount:
          type: integer
      required: [id, categoryId, categoryName, totalAmount, availablePlayers, perPlayerAmount, weekStartDate, createdBy, createdAt, playerFees, paidCount, unpaidCount]

    PlayerFee:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryFeeId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        playerName:
          type: string
        status:
          type: string
          enum: [pending, paid]
        paidAt:
          type: string
          format: date-time
          nullable: true
      required: [id, categoryFeeId, userId, playerName, status, paidAt]

    CreateCategoryFeeRequest:
      type: object
      properties:
        categoryId:
          type: string
        totalAmount:
          type: number
          minimum: 0.01
        availablePlayers:
          type: integer
          minimum: 1
      required: [categoryId, totalAmount, availablePlayers]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT